+++ neuronpp/examples/combe2018_iv_curves.py
@@ -1,7 +1,7 @@
 from neuron import h
 import numpy as np
 import matplotlib.pyplot as plt
-from neuronpp.cells.combe2018_cell_python_from_scratch import Combe2018Cell
+from neuronpp.cells.combe2018 import Combe2018Cell
 from neuronpp.utils.simulation import Simulation
 from neuronpp.utils.record import Record
 from neuronpp.utils.iclamp import IClamp
